Split out the upload page view from the base classes, since it is not a base class anymore. Also tidied docstrings
