+++ timetable_solutions/interfaces/data_upload/views/__init__.py
@@ -1,5 +1,7 @@
-# Upload views
-from .base_classes import UploadPage
+"""
+All views relating to the data upload page
+"""
+from .data_upload_page import UploadPage
 from .upload_views import (PupilListUpload, TeacherListUpload, ClassroomListUpload, TimetableStructureUpload,
                            UnsolvedClassUpload, FixedClassUpload)
 from .reset_views import (PupilListReset, TeacherListReset, ClassroomListReset, TimetableStructureReset,
